moved dialog component and added styles
